+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Gerapy Pyppeteer Changelog
 
+## 0.0.10 (2020-08-01)
+
+### Features
+
+* Change the priority of `request.meta.get('proxy')` and `pyppeteer_meta.get('proxy')`
+* Add `pretend` attribute for `PyppeteerRequest`, which can override `GERAPY_PYPPETEER_PRETEND`
+ 
 ## 0.0.9 (2020-07-31)
 
 ### Features